feat: renamed apiService.tsx to service.tsx
